fix(wrapped-keys): LIT-3125 - Remove inadvertent imports from `index.ts` and eliminate circular type imports
