+++ packages/wrapped-keys/src/lib/service-client/index.ts
@@ -1,9 +1,3 @@
 import { fetchPrivateKeyMetadata, storePrivateKeyMetadata } from './client';
-import { StoredKeyMetadata, SupportedNetworks } from './types';
 
-export {
-  fetchPrivateKeyMetadata,
-  storePrivateKeyMetadata,
-  StoredKeyMetadata,
-  SupportedNetworks,
-};
+export { fetchPrivateKeyMetadata, storePrivateKeyMetadata };
