+++ packages/wrapped-keys/src/lib/types.ts
@@ -4,8 +4,6 @@ import {
   SessionSigsMap,
 } from '@lit-protocol/types';
 
-import { StoredKeyMetadata } from '../index';
-
 /** @typedef Network
  * The network type that the wrapped key will be used on.
  */
@@ -34,6 +32,25 @@ export interface ApiParamsSupportedNetworks {
  */
 
 export type GetEncryptedKeyMetadataParams = BaseApiParams;
+
+/** Metadata for a key that has been stored, encrypted, on the wrapped keys backend service
+ *
+ * @property { string } ciphertext The base64 encoded, salted & encrypted private key
+ * @property { string } dataToEncryptHash SHA-256 of the ciphertext
+ * @property { string } publicKey The public key of the encrypted private key
+ * @property { string } pkpAddress The LIT PKP address that is associated with the encrypted private key
+ * @property { string } keyType The type of key that was encrypted -- e.g. ed25519, K256, etc.
+ * @property { LIT_NETWORKS_KEYS } litNetwork The LIT network that the client who stored the key was connected to
+ */
+export interface StoredKeyMetadata {
+  ciphertext: string;
+  dataToEncryptHash: string;
+  publicKey: string;
+  pkpAddress: string;
+  keyType: string;
+  litNetwork: LIT_NETWORKS_KEYS;
+}
+
 /** Fetching a previously persisted key's metadata only requires valid pkpSessionSigs and a LIT Node Client instance configured for the appropriate network.
  *
  * @typedef StoreEncryptedKeyMetadataParams
